backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
groq==0.4.2
python-multipart==0.0.20
python-dotenv==1.0.0
gunicorn==21.2.0
